Latest Release on PATREON - UPD telcoma proto

+++ lib/subghz/protocols/telcoma_edge.h
@@ -36,7 +36,7 @@ void* subghz_protocol_decoder_telcoma_edge_alloc(SubGhzEnvironment* environment)
 void subghz_protocol_decoder_telcoma_edge_free(void* context);
 void subghz_protocol_decoder_telcoma_edge_reset(void* context);
 void subghz_protocol_decoder_telcoma_edge_feed(void* context, bool level, uint32_t duration);
-uint8_t subghz_protocol_decoder_telcoma_edge_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_telcoma_edge_get_hash_data(void* context);
 SubGhzProtocolStatus subghz_protocol_decoder_telcoma_edge_serialize(
     void* context,
     FlipperFormat* flipper_format,
